+++ lib/active_mocker/mock/compatibility/queries/ar52.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+module ActiveMocker
+  module Queries
+    module AR52
+      extend LateInclusion::Extension
+
+      private
+
+      def check_for_limit_scope!
+        # noop
+      end
+    end
+  end
+end